requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain==0.1.5
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pillow==10.0.1